hera/requirements.txt
requests
PyJWT
pandas
python-docx
PyPDF2
markdownify
openpyxl
tabulate